requirements1.txt
apturl==0.5.2
asgiref==3.8.1
astroid==3.3.5
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
cloud-init==24.4.1
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
dill==0.3.9
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==4.2.17
django-admin-interface==0.29.2
django-adminlte3==0.1.6
django-axes==7.0.2
django-colorfield==0.11.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-guardian==2.4.0
django-jet-reboot==1.3.10
django-mssql-backend==2.8.1
djangorestframework==3.15.2
duplicity==0.8.21
et_xmlfile==2.0.0
fasteners==0.14.1
future==0.18.2
fuzzywuzzy==0.18.0
graphviz==0.20.3
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
isort==5.13.2
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.7.0
monotonic==1.6
more-itertools==8.10.0
mssql-django==1.5
netifaces==0.11.0
numpy==2.2.4
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.5
packaging==21.3
pandas==2.2.3
paramiko==2.9.3
pexpect==4.8.0
pillow==11.0.0
platformdirs==4.3.6
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylint==3.3.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pyodbc==5.2.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-magic==0.4.24
python-slugify==8.0.4
pytz==2024.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.8.2
sqlparse==0.5.2
ssh-import-id==5.11
systemd-python==234
terminator==2.1.1
text-unidecode==1.3
tomli==2.2.1
tomlkit==0.13.2
Twisted==22.1.0
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
xlrd==2.0.1
zipp==1.0.0
zope.interface==5.4.0
